review-kluebert-preprocessing/requirements.txt
functions-framework==3.*
python-mecab-ko==1.3.5
python-mecab-ko-dic==2.1.1.post2
kss==6.0.4
google-cloud-pubsub==2.21.5
pandas